move obcq to pruning.sparsegpt
