+++ src/llmcompressor/modifiers/pruning/__init__.py
@@ -2,4 +2,5 @@
 
 from .constant import *
 from .magnitude import *
+from .sparsegpt import *
 from .wanda import *
