+++ tests/llmcompressor/transformers/finetune/test_finetune_oneshot_with_modifier.py
@@ -22,7 +22,7 @@ class TestOneshotWithModifierObject(unittest.TestCase):
         self.output = Path("./finetune_output")
 
     def test_oneshot_with_modifier_object(self):
-        from llmcompressor.modifiers.obcq.base import SparseGPTModifier
+        from llmcompressor.modifiers.pruning.sparsegpt.base import SparseGPTModifier
         from llmcompressor.transformers import oneshot
 
         recipe_str = [
