+++ tests/llmcompressor/transformers/obcq/recipes/additional_sparsity_with_quant.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     QuantizationModifier:
       config_groups:
         group_0:
