+++ tests/llmcompressor/transformers/obcq/recipes/quant_and_sparse.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     SmoothQuantModifier:
       smoothing_strength: 0.5
       mappings: [
