+++ tests/llmcompressor/transformers/obcq/recipes/sparse.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     SparseGPTModifier:
       sparsity: 0.3
       block_size: 128
