+++ tests/llmcompressor/transformers/obcq/recipes/sparse_with_mask_structure.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     SparseGPTModifier:
       sparsity: 0.5
       block_size: 128
