+++ tests/llmcompressor/transformers/obcq/recipes/test_tiny2.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     SparseGPTModifier:
       sparsity: 0.5
       block_size: 128
