+++ tests/llmcompressor/transformers/oneshot/oneshot_configs/recipes/recipe.yaml
@@ -1,5 +1,5 @@
 test_stage:
-  obcq_modifiers:
+  modifiers:
     SparseGPTModifier:
       sparsity: 0.5
       block_size: 128
