+++ tests/llmcompressor/transformers/oneshot/oneshot_configs/tiny_stories_conf1.yaml
@@ -5,7 +5,7 @@ model: "Xenova/llama2.c-stories15M"
 dataset: open_platypus
 recipe: |
   test_stage:
-    obcq_modifiers:
+    modifiers:
       SparseGPTModifier:
         sparsity: 0.5
         block_size: 128
