+++ tests/llmcompressor/transformers/oneshot/oneshot_configs/tiny_stories_conf4.yaml
@@ -6,7 +6,7 @@ dataset: "gsm8k"
 dataset_config_name: "main"
 recipe: |
   test_stage:
-    obcq_modifiers:
+    modifiers:
       SparseGPTModifier:
         sparsity: 0.5
         block_size: 128
